+++ languages/cpp/queries/injections.scm
@@ -1,3 +1,5 @@
+((comment) @injection.content
+ (#set! injection.language "comment"))
 (raw_string_literal
   delimiter: (raw_string_delimiter) @injection.language
   (raw_string_content) @injection.content)