+++ languages/glsl/queries/injections.scm
@@ -1,4 +1,5 @@
-; inherits: c
+((comment) @injection.content
+ (#set! injection.language "comment"))
 
 ((preproc_arg) @injection.content
  (#set! injection.language "glsl"))