+++ src/theme/vendored/mod.rs
@@ -4,32 +4,25 @@ pub const ACME: &str = include_str!("./data/acme.toml");
 pub const ADWAITA_DARK: &str = include_str!("./data/adwaita-dark.toml");
 pub const AMBERWOOD: &str = include_str!("./data/amberwood.toml");
 pub const AO: &str = include_str!("./data/ao.toml");
-pub const AUTUMN_NIGHT: &str = include_str!("./data/autumn_night.toml");
-pub const AUTUMN: &str = include_str!("./data/autumn.toml");
 pub const AYU_DARK: &str = include_str!("./data/ayu_dark.toml");
-pub const AYU_EVOLVE: &str = include_str!("./data/ayu_evolve.toml");
 pub const AYU_LIGHT: &str = include_str!("./data/ayu_light.toml");
 pub const AYU_MIRAGE: &str = include_str!("./data/ayu_mirage.toml");
 pub const BASE16_DEFAULT_DARK: &str = include_str!("./data/base16_default_dark.toml");
 pub const BASE16_DEFAULT_LIGHT: &str = include_str!("./data/base16_default_light.toml");
 pub const BASE16_TERMINAL: &str = include_str!("./data/base16_terminal.toml");
 pub const BASE16_TRANSPARENT: &str = include_str!("./data/base16_transparent.toml");
-pub const BOGSTER_LIGHT: &str = include_str!("./data/bogster_light.toml");
 pub const BOGSTER: &str = include_str!("./data/bogster.toml");
+pub const BOGSTER_LIGHT: &str = include_str!("./data/bogster_light.toml");
 pub const BOO_BERRY: &str = include_str!("./data/boo_berry.toml");
-pub const CATPPUCCIN_FRAPPE: &str = include_str!("./data/catppuccin_frappe.toml");
-pub const CATPPUCCIN_LATTE: &str = include_str!("./data/catppuccin_latte.toml");
-pub const CATPPUCCIN_MACCHIATO: &str = include_str!("./data/catppuccin_macchiato.toml");
 pub const CATPPUCCIN_MOCHA: &str = include_str!("./data/catppuccin_mocha.toml");
 pub const CURZON: &str = include_str!("./data/curzon.toml");
 pub const CYAN_LIGHT: &str = include_str!("./data/cyan_light.toml");
-pub const DARCULA_SOLID: &str = include_str!("./data/darcula-solid.toml");
 pub const DARCULA: &str = include_str!("./data/darcula.toml");
 pub const DARK_HIGH_CONTRAST: &str = include_str!("./data/dark_high_contrast.toml");
 pub const DARK_PLUS: &str = include_str!("./data/dark_plus.toml");
 pub const DOOM_ACARIO_DARK: &str = include_str!("./data/doom_acario_dark.toml");
-pub const DRACULA_AT_NIGHT: &str = include_str!("./data/dracula_at_night.toml");
 pub const DRACULA: &str = include_str!("./data/dracula.toml");
+pub const DRACULA_AT_NIGHT: &str = include_str!("./data/dracula_at_night.toml");
 pub const EMACS: &str = include_str!("./data/emacs.toml");
 pub const EVERBLUSH: &str = include_str!("./data/everblush.toml");
 pub const EVERFOREST_DARK: &str = include_str!("./data/everforest_dark.toml");
@@ -37,101 +30,51 @@ pub const EVERFOREST_LIGHT: &str = include_str!("./data/everforest_light.toml");
 pub const FERRA: &str = include_str!("./data/ferra.toml");
 pub const FLATWHITE: &str = include_str!("./data/flatwhite.toml");
 pub const FLEET_DARK: &str = include_str!("./data/fleet_dark.toml");
-pub const FLEXOKI_DARK: &str = include_str!("./data/flexoki_dark.toml");
 pub const FLEXOKI_LIGHT: &str = include_str!("./data/flexoki_light.toml");
-pub const GITHUB_DARK_COLORBLIND: &str = include_str!(
-    "./data/github_dark_colorblind.toml"
-);
-pub const GITHUB_DARK_DIMMED: &str = include_str!("./data/github_dark_dimmed.toml");
-pub const GITHUB_DARK_HIGH_CONTRAST: &str = include_str!(
-    "./data/github_dark_high_contrast.toml"
-);
-pub const GITHUB_DARK_TRITANOPIA: &str = include_str!(
-    "./data/github_dark_tritanopia.toml"
-);
 pub const GITHUB_DARK: &str = include_str!("./data/github_dark.toml");
-pub const GITHUB_LIGHT_COLORBLIND: &str = include_str!(
-    "./data/github_light_colorblind.toml"
-);
-pub const GITHUB_LIGHT_HIGH_CONTRAST: &str = include_str!(
-    "./data/github_light_high_contrast.toml"
-);
-pub const GITHUB_LIGHT_TRITANOPIA: &str = include_str!(
-    "./data/github_light_tritanopia.toml"
-);
 pub const GITHUB_LIGHT: &str = include_str!("./data/github_light.toml");
 pub const GRUBER_DARKER: &str = include_str!("./data/gruber-darker.toml");
-pub const GRUVBOX_DARK_HARD: &str = include_str!("./data/gruvbox_dark_hard.toml");
-pub const GRUVBOX_DARK_SOFT: &str = include_str!("./data/gruvbox_dark_soft.toml");
-pub const GRUVBOX_LIGHT_HARD: &str = include_str!("./data/gruvbox_light_hard.toml");
-pub const GRUVBOX_LIGHT_SOFT: &str = include_str!("./data/gruvbox_light_soft.toml");
-pub const GRUVBOX_LIGHT: &str = include_str!("./data/gruvbox_light.toml");
 pub const GRUVBOX: &str = include_str!("./data/gruvbox.toml");
 pub const HEISENBERG: &str = include_str!("./data/heisenberg.toml");
-pub const HEX_LAVENDER: &str = include_str!("./data/hex_lavender.toml");
 pub const HEX_STEEL: &str = include_str!("./data/hex_steel.toml");
-pub const HEX_TOXIC: &str = include_str!("./data/hex_toxic.toml");
 pub const HORIZON_DARK: &str = include_str!("./data/horizon-dark.toml");
 pub const ICEBERG_DARK: &str = include_str!("./data/iceberg-dark.toml");
-pub const ICEBERG_LIGHT: &str = include_str!("./data/iceberg-light.toml");
 pub const INGRID: &str = include_str!("./data/ingrid.toml");
 pub const IROASETA: &str = include_str!("./data/iroaseta.toml");
 pub const JELLYBEANS: &str = include_str!("./data/jellybeans.toml");
 pub const JETBRAINS_DARK: &str = include_str!("./data/jetbrains_dark.toml");
-pub const KANAGAWA_DRAGON: &str = include_str!("./data/kanagawa-dragon.toml");
 pub const KANAGAWA: &str = include_str!("./data/kanagawa.toml");
 pub const KAOLIN_DARK: &str = include_str!("./data/kaolin-dark.toml");
-pub const KAOLIN_LIGHT: &str = include_str!("./data/kaolin-light.toml");
-pub const KAOLIN_VALLEY_DARK: &str = include_str!("./data/kaolin-valley-dark.toml");
-pub const MATERIAL_DARKER: &str = include_str!("./data/material_darker.toml");
 pub const MATERIAL_DEEP_OCEAN: &str = include_str!("./data/material_deep_ocean.toml");
-pub const MATERIAL_OCEANIC: &str = include_str!("./data/material_oceanic.toml");
-pub const MATERIAL_PALENIGHT: &str = include_str!("./data/material_palenight.toml");
 pub const MELIORA: &str = include_str!("./data/meliora.toml");
 pub const MELLOW: &str = include_str!("./data/mellow.toml");
 pub const MERIONETTE: &str = include_str!("./data/merionette.toml");
-pub const MODUS_OPERANDI_DEUTERANOPIA: &str = include_str!(
-    "./data/modus_operandi_deuteranopia.toml"
-);
-pub const MODUS_OPERANDI_TINTED: &str = include_str!(
-    "./data/modus_operandi_tinted.toml"
-);
-pub const MODUS_OPERANDI_TRITANOPIA: &str = include_str!(
-    "./data/modus_operandi_tritanopia.toml"
-);
 pub const MODUS_OPERANDI: &str = include_str!("./data/modus_operandi.toml");
-pub const MOLOKAI: &str = include_str!("./data/molokai.toml");
-pub const MONOKAI_AQUA: &str = include_str!("./data/monokai_aqua.toml");
+pub const MONOKAI: &str = include_str!("./data/monokai.toml");
+pub const MONOKAI_PRO: &str = include_str!("./data/monokai_pro.toml");
 pub const MONOKAI_PRO_MACHINE: &str = include_str!("./data/monokai_pro_machine.toml");
 pub const MONOKAI_PRO_OCTAGON: &str = include_str!("./data/monokai_pro_octagon.toml");
 pub const MONOKAI_PRO_RISTRETTO: &str = include_str!(
     "./data/monokai_pro_ristretto.toml"
 );
 pub const MONOKAI_PRO_SPECTRUM: &str = include_str!("./data/monokai_pro_spectrum.toml");
-pub const MONOKAI_PRO: &str = include_str!("./data/monokai_pro.toml");
 pub const MONOKAI_SODA: &str = include_str!("./data/monokai_soda.toml");
-pub const MONOKAI: &str = include_str!("./data/monokai.toml");
 pub const NAYSAYER: &str = include_str!("./data/naysayer.toml");
 pub const NEW_MOON: &str = include_str!("./data/new_moon.toml");
 pub const NIGHT_OWL: &str = include_str!("./data/night_owl.toml");
 pub const NIGHTFOX: &str = include_str!("./data/nightfox.toml");
-pub const NOCTIS_BORDO: &str = include_str!("./data/noctis_bordo.toml");
 pub const NOCTIS: &str = include_str!("./data/noctis.toml");
-pub const NORD_LIGHT: &str = include_str!("./data/nord_light.toml");
-pub const NORD_NIGHT: &str = include_str!("./data/nord-night.toml");
+pub const NOCTIS_BORDO: &str = include_str!("./data/noctis_bordo.toml");
 pub const NORD: &str = include_str!("./data/nord.toml");
+pub const NORD_LIGHT: &str = include_str!("./data/nord_light.toml");
 pub const ONEDARK: &str = include_str!("./data/onedark.toml");
 pub const ONEDARKER: &str = include_str!("./data/onedarker.toml");
 pub const ONELIGHT: &str = include_str!("./data/onelight.toml");
-pub const PAPERCOLOR_DARK: &str = include_str!("./data/papercolor-dark.toml");
 pub const PAPERCOLOR_LIGHT: &str = include_str!("./data/papercolor-light.toml");
 pub const PENUMBRA_PLUS: &str = include_str!("./data/penumbra-plus.toml");
-pub const POIMANDRES_STORM: &str = include_str!("./data/poimandres_storm.toml");
 pub const POIMANDRES: &str = include_str!("./data/poimandres.toml");
 pub const POP_DARK: &str = include_str!("./data/pop-dark.toml");
 pub const RASMUS: &str = include_str!("./data/rasmus.toml");
-pub const ROSE_PINE_DAWN: &str = include_str!("./data/rose_pine_dawn.toml");
-pub const ROSE_PINE_MOON: &str = include_str!("./data/rose_pine_moon.toml");
 pub const ROSE_PINE: &str = include_str!("./data/rose_pine.toml");
 pub const SERIKA_DARK: &str = include_str!("./data/serika-dark.toml");
 pub const SERIKA_LIGHT: &str = include_str!("./data/serika-light.toml");
@@ -142,10 +85,6 @@ pub const SONOKAI: &str = include_str!("./data/sonokai.toml");
 pub const SPACEBONES_LIGHT: &str = include_str!("./data/spacebones_light.toml");
 pub const STARLIGHT: &str = include_str!("./data/starlight.toml");
 pub const TERM16_DARK: &str = include_str!("./data/term16_dark.toml");
-pub const TERM16_LIGHT: &str = include_str!("./data/term16_light.toml");
-pub const TOKYONIGHT_DAY: &str = include_str!("./data/tokyonight_day.toml");
-pub const TOKYONIGHT_MOON: &str = include_str!("./data/tokyonight_moon.toml");
-pub const TOKYONIGHT_STORM: &str = include_str!("./data/tokyonight_storm.toml");
 pub const TOKYONIGHT: &str = include_str!("./data/tokyonight.toml");
 pub const TTOX: &str = include_str!("./data/ttox.toml");
 pub const VARUA: &str = include_str!("./data/varua.toml");
@@ -155,5 +94,464 @@ pub const VIM_DARK_HIGH_CONTRAST: &str = include_str!(
 pub const VOXED: &str = include_str!("./data/voxed.toml");
 pub const YELLOWED: &str = include_str!("./data/yellowed.toml");
 pub const ZED_ONEDARK: &str = include_str!("./data/zed_onedark.toml");
-pub const ZED_ONELIGHT: &str = include_str!("./data/zed_onelight.toml");
 pub const ZENBURN: &str = include_str!("./data/zenburn.toml");
+#[cfg(test)]
+#[allow(non_snake_case)]
+mod tests {
+    use crate::theme::*;
+    #[test]
+    fn ACME() {
+        let data = include_str!("./data/acme.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn ADWAITA_DARK() {
+        let data = include_str!("./data/adwaita-dark.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn AMBERWOOD() {
+        let data = include_str!("./data/amberwood.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn AO() {
+        let data = include_str!("./data/ao.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn AYU_DARK() {
+        let data = include_str!("./data/ayu_dark.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn AYU_LIGHT() {
+        let data = include_str!("./data/ayu_light.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn AYU_MIRAGE() {
+        let data = include_str!("./data/ayu_mirage.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn BASE16_DEFAULT_DARK() {
+        let data = include_str!("./data/base16_default_dark.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn BASE16_DEFAULT_LIGHT() {
+        let data = include_str!("./data/base16_default_light.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn BASE16_TERMINAL() {
+        let data = include_str!("./data/base16_terminal.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn BASE16_TRANSPARENT() {
+        let data = include_str!("./data/base16_transparent.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn BOGSTER() {
+        let data = include_str!("./data/bogster.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn BOGSTER_LIGHT() {
+        let data = include_str!("./data/bogster_light.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn BOO_BERRY() {
+        let data = include_str!("./data/boo_berry.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn CATPPUCCIN_MOCHA() {
+        let data = include_str!("./data/catppuccin_mocha.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn CURZON() {
+        let data = include_str!("./data/curzon.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn CYAN_LIGHT() {
+        let data = include_str!("./data/cyan_light.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn DARCULA() {
+        let data = include_str!("./data/darcula.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn DARK_HIGH_CONTRAST() {
+        let data = include_str!("./data/dark_high_contrast.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn DARK_PLUS() {
+        let data = include_str!("./data/dark_plus.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn DOOM_ACARIO_DARK() {
+        let data = include_str!("./data/doom_acario_dark.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn DRACULA() {
+        let data = include_str!("./data/dracula.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn DRACULA_AT_NIGHT() {
+        let data = include_str!("./data/dracula_at_night.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn EMACS() {
+        let data = include_str!("./data/emacs.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn EVERBLUSH() {
+        let data = include_str!("./data/everblush.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn EVERFOREST_DARK() {
+        let data = include_str!("./data/everforest_dark.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn EVERFOREST_LIGHT() {
+        let data = include_str!("./data/everforest_light.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn FERRA() {
+        let data = include_str!("./data/ferra.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn FLATWHITE() {
+        let data = include_str!("./data/flatwhite.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn FLEET_DARK() {
+        let data = include_str!("./data/fleet_dark.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn FLEXOKI_LIGHT() {
+        let data = include_str!("./data/flexoki_light.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn GITHUB_DARK() {
+        let data = include_str!("./data/github_dark.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn GITHUB_LIGHT() {
+        let data = include_str!("./data/github_light.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn GRUBER_DARKER() {
+        let data = include_str!("./data/gruber-darker.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn GRUVBOX() {
+        let data = include_str!("./data/gruvbox.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn HEISENBERG() {
+        let data = include_str!("./data/heisenberg.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn HEX_STEEL() {
+        let data = include_str!("./data/hex_steel.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn HORIZON_DARK() {
+        let data = include_str!("./data/horizon-dark.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn ICEBERG_DARK() {
+        let data = include_str!("./data/iceberg-dark.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn INGRID() {
+        let data = include_str!("./data/ingrid.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn IROASETA() {
+        let data = include_str!("./data/iroaseta.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn JELLYBEANS() {
+        let data = include_str!("./data/jellybeans.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn JETBRAINS_DARK() {
+        let data = include_str!("./data/jetbrains_dark.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn KANAGAWA() {
+        let data = include_str!("./data/kanagawa.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn KAOLIN_DARK() {
+        let data = include_str!("./data/kaolin-dark.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn MATERIAL_DEEP_OCEAN() {
+        let data = include_str!("./data/material_deep_ocean.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn MELIORA() {
+        let data = include_str!("./data/meliora.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn MELLOW() {
+        let data = include_str!("./data/mellow.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn MERIONETTE() {
+        let data = include_str!("./data/merionette.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn MODUS_OPERANDI() {
+        let data = include_str!("./data/modus_operandi.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn MONOKAI() {
+        let data = include_str!("./data/monokai.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn MONOKAI_PRO() {
+        let data = include_str!("./data/monokai_pro.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn MONOKAI_PRO_MACHINE() {
+        let data = include_str!("./data/monokai_pro_machine.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn MONOKAI_PRO_OCTAGON() {
+        let data = include_str!("./data/monokai_pro_octagon.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn MONOKAI_PRO_RISTRETTO() {
+        let data = include_str!("./data/monokai_pro_ristretto.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn MONOKAI_PRO_SPECTRUM() {
+        let data = include_str!("./data/monokai_pro_spectrum.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn MONOKAI_SODA() {
+        let data = include_str!("./data/monokai_soda.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn NAYSAYER() {
+        let data = include_str!("./data/naysayer.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn NEW_MOON() {
+        let data = include_str!("./data/new_moon.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn NIGHT_OWL() {
+        let data = include_str!("./data/night_owl.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn NIGHTFOX() {
+        let data = include_str!("./data/nightfox.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn NOCTIS() {
+        let data = include_str!("./data/noctis.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn NOCTIS_BORDO() {
+        let data = include_str!("./data/noctis_bordo.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn NORD() {
+        let data = include_str!("./data/nord.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn NORD_LIGHT() {
+        let data = include_str!("./data/nord_light.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn ONEDARK() {
+        let data = include_str!("./data/onedark.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn ONEDARKER() {
+        let data = include_str!("./data/onedarker.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn ONELIGHT() {
+        let data = include_str!("./data/onelight.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn PAPERCOLOR_LIGHT() {
+        let data = include_str!("./data/papercolor-light.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn PENUMBRA_PLUS() {
+        let data = include_str!("./data/penumbra-plus.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn POIMANDRES() {
+        let data = include_str!("./data/poimandres.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn POP_DARK() {
+        let data = include_str!("./data/pop-dark.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn RASMUS() {
+        let data = include_str!("./data/rasmus.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn ROSE_PINE() {
+        let data = include_str!("./data/rose_pine.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn SERIKA_DARK() {
+        let data = include_str!("./data/serika-dark.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn SERIKA_LIGHT() {
+        let data = include_str!("./data/serika-light.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn SNAZZY() {
+        let data = include_str!("./data/snazzy.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn SOLARIZED_DARK() {
+        let data = include_str!("./data/solarized_dark.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn SOLARIZED_LIGHT() {
+        let data = include_str!("./data/solarized_light.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn SONOKAI() {
+        let data = include_str!("./data/sonokai.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn SPACEBONES_LIGHT() {
+        let data = include_str!("./data/spacebones_light.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn STARLIGHT() {
+        let data = include_str!("./data/starlight.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn TERM16_DARK() {
+        let data = include_str!("./data/term16_dark.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn TOKYONIGHT() {
+        let data = include_str!("./data/tokyonight.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn TTOX() {
+        let data = include_str!("./data/ttox.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn VARUA() {
+        let data = include_str!("./data/varua.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn VIM_DARK_HIGH_CONTRAST() {
+        let data = include_str!("./data/vim_dark_high_contrast.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn VOXED() {
+        let data = include_str!("./data/voxed.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn YELLOWED() {
+        let data = include_str!("./data/yellowed.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn ZED_ONEDARK() {
+        let data = include_str!("./data/zed_onedark.toml");
+        Theme::from_helix(data).unwrap();
+    }
+    #[test]
+    fn ZENBURN() {
+        let data = include_str!("./data/zenburn.toml");
+        Theme::from_helix(data).unwrap();
+    }
+}