+++ tests/snippets.rs
@@ -1,8 +1,9 @@
 use std::fs::File;
 use std::io::Write;
 
+use formatter::ThemedHtml;
 use inkjet::*;
-use inkjet::formatter::Html;
+use inkjet::theme::*;
 
 #[test]
 fn highlight_all_snippets() {
@@ -11,6 +12,12 @@ fn highlight_all_snippets() {
 
     let mut hili = Highlighter::new();
 
+    let theme = Theme::from_helix(vendored::ADWAITA_DARK)
+        .map(ThemedHtml::new)
+        .unwrap();
+
+    eprintln!("{:#?}", theme);
+
     for lang in Language::ALL_LANGS {
         let source = std::fs::read_to_string(
             format!("./tests/samples/{lang:?}")
@@ -18,20 +25,15 @@ fn highlight_all_snippets() {
         
         let _ = write!(
             &mut out,
-            "<h2>{lang:?}</h2><pre>"
+            "<h2>{lang:?}</h2><br>"
         );
 
         hili.highlight_to_writer(
             *lang,
-            &Html,
+            &theme,
             source,
             &mut out
         ).unwrap();
-
-        let _ = write!(
-            &mut out,
-            "</pre>"
-        );
     }
 }